Beautifying log pages (2/2)
Log page is more "beautiful", the latest messages now show channel as a lighter color and the message and sender as a darker color. The table is also now unified and the more important columns are bigger. The text-decorations of links has also been removed, unless you mouse over it.